% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_function.R
\name{add}
\alias{add}
\title{Basic functions}
\usage{
add(a, b)
}
\arguments{
\item{a}{input 1 (number)}

\item{b}{input 2 (number)}
}
\description{
This is a simple add function which takes in two variables
and returns the sum of the two variables.
}
\details{
To add tests for this file run \code{use_test()} with this file
open.
}
