% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erropadrao.R
\name{erropadrao}
\alias{erropadrao}
\title{Ferramenta para calcular o erro padrão de um vetor}
\usage{
erropadrao(x)
}
\arguments{
\item{x}{o vetor a ser analizado}
}
\description{
Esta função permite calcular o erro padrão a partir de o desvio padrão e o tamanho da amostra
}
\examples{
cat_function()
}
\keyword{erropadrão,}
\keyword{error}
\keyword{standard}

