% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1}
\alias{f1}
\title{This is the f1 function}
\usage{
f1(text = "default")
}
\arguments{
\item{text}{(character) the text to paste with f1. Default = "default"}
}
\value{
character vector of length 1
}
\description{
This is the f1 function
}
\examples{
f <- f1(text = "default")
}
