% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{make a map from data using state number and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer that represents state number}

\item{year}{an integer that represents a year}
}
\value{
A map of accidents for the given state during the given year
}
\description{
make a map from data using state number and year
}
\examples{
\dontrun{fars_map_state(4, 2022)}
}
