% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{load a file with filename}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a file name of a .csv file}
}
\value{
A data.frame

dataframe in environment and tibble in console
}
\description{
load a file with filename
}
\note{
file name must exist in your drive, if not with stop with error.
}
\examples{
\dontrun{fars_read(User/Document/file.csv)}
}
