% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{summarizes data by month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of integers that represents years}
}
\value{
returns a data.frame summarizing n by month and year
}
\description{
summarizes data by month and year
}
\examples{
\dontrun{fars_summarize_years(2022)}
\dontrun{fars_summarize_years(c(2020, 2021, 2022))}
}
