% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Binder into a Factor
Binds Two vectors into a Factor}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor/character}

\item{b}{factor/character}
}
\value{
factor
}
\description{
Binder into a Factor
Binds Two vectors into a Factor
}
\examples{
char_a <- c("male", "female", "neutral")
fact_b <- factor(c("Pilani", "Hyderabad", "Goa", "Dubai"))
fbind(char_a,fact_b)
}
