% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_missing.R
\name{fix_missing}
\alias{fix_missing}
\title{Valeurs à considérer comme manquantes}
\usage{
fix_missing(x, na.value)
}
\arguments{
\item{x}{un vecteur}

\item{na.value}{une valeur à considérer comme manquantes}
}
\value{
vecteur avec des valeurs manquantes à la place des occurences de na.value
}
\description{
Valeurs à considérer comme manquantes
}
\examples{
valeurs <- c('1','9','3','1')
fix_missing(valeurs,9)
#' @encoding UTF-8
}
