% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_missing_df.R
\encoding{UTF-8}
\name{fix_missing_df}
\alias{fix_missing_df}
\title{Fixmission pourl es data frame}
\usage{
fix_missing_df(x, na.value)
}
\arguments{
\item{x}{data framee en ettre}

\item{na.value}{valeur à coder en manquante}
}
\value{
data frame avec les valeurs manquantes
}
\description{
Fixmission pourl es data frame
}
\examples{
df <- data.frame(v1 = 1:10,v2=8:17,v3=18:27)
fix_missing_df(df,9)
}
