% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello there!}
\usage{
hello(arg)
}
\arguments{
\item{arg}{An argument. Can be whatever.}
}
\value{
a character object, chr "hello"
}
\description{
This function returns the character object "hello".
}
