% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment.R
\name{increment}
\alias{increment}
\title{Increment}
\usage{
increment(value)
}
\arguments{
\item{value}{This could be anything when passes. Lets check that the desired output
is recieved for a range of input types.}
}
\description{
Takes in a variable value and increments it by 1.
}
