% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{create a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer that represents a year}
}
\value{
A generated file name for a given year

character string in form of "accident_2021.csv.bz2"
}
\description{
create a file name
}
\examples{
\dontrun{make_filename(2021)}
}
