\name{packagetest-package}
\alias{packagetest-package}
\alias{packagetest}
\docType{package}
\title{
\packageTitle{packagetest}
}
\description{
\packageDescription{packagetest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{packagetest}
\packageIndices{packagetest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{packagetest}

Maintainer: \packageMaintainer{packagetest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
