% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotASEMetrics.R
\name{plotASEMetrics}
\alias{plotASEMetrics}
\title{Plot ASE metrics generated by Gen.input.}
\usage{
plotASEMetrics(input, individuals = NULL, genes = NULL, variants = NULL)
}
\arguments{
\item{input}{The output of the Gen.input function}

\item{individuals}{The individual(s) to restrict the plots to.}

\item{genes}{The gene(s) to restrict the plots to.}

\item{variants}{The variant(s) to restrict the plots to.}
}
\description{
This function takes the output of Gen.input and produces various plots regarding allele specific expression in the dataset.
}
\examples{
#' #Plot the output of Gen.input stored in aseDat
plotASEMetrics(aseDat)
                    
#Annotating the input files using a previous human gene set
Gen.input("path_to_ASE_file/ASEfile.txt.gz",
                    "path_to_legend_file/file.legendfile.gz",
                   "path_to_haplotypes_file/hapfile.hap.gz",
                    "path_to_samples_file/samples.txt", ensembl_version=78)

#Annotating the input files using the most recent mouse gene set
Gen.input("path_to_ASE_file/ASEfile.txt.gz",
                    "path_to_legend_file/file.legendfile.gz",
                  "path_to_haplotypes_file/hapfile.hap.gz",
                    "path_to_samples_file/samples.txt",
                    "mmusculus")
}
