% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Hello}
\alias{Hello}
\title{Hello there!}
\usage{
Hello(arg)
}
\arguments{
\item{arg}{An argument. Can be whatever.}
}
\value{
a character object, chr "hello"
}
\description{
This function prints the string 'hello' when given an argument. It returns the character object generated by the print() function.
}

