% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageTEST-package.R
\docType{package}
\name{PackageTEST-package}
\alias{PackageTEST}
\alias{PackageTEST-package}
\title{PackageTEST: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NefrolithiasisDrHorse/PackageTEST}
  \item Report bugs at \url{https://github.com/NefrolithiasisDrHorse/PackageTEST/issues}
}

}
\author{
\strong{Maintainer}: Peter Hjort \email{first.last@example.com}

}
\keyword{internal}
